separating default libmr from new weibull functionality
